re-enabled dynamic version information
- added mdpow.version that can imported without side effects
- setup.py is again using the get_version() so that there is only
  a single place where the version information needs to be changed
